+++ src/fr/maxlego08/menu/ZInventoryManager.java
@@ -265,7 +265,7 @@ public class ZInventoryManager extends ZUtils implements InventoryManager {
         // Load Permissible before
         buttonManager.registerPermissible(new PlaceholderPermissibleLoader(buttonManager));
         buttonManager.registerPermissible(new PermissionPermissibleLoader(buttonManager));
-        buttonManager.registerPermissible(new ItemPermissibleLoader(buttonManager));
+        buttonManager.registerPermissible(new ItemPermissibleLoader(this.plugin));
         buttonManager.registerPermissible(new RegexPermissibleLoader(buttonManager));
         buttonManager.registerPermissible(new PlayerNamePermissibleLoader(buttonManager));
         if (this.plugin.isEnable(Plugins.JOBS)) {
@@ -701,5 +701,12 @@ public class ZInventoryManager extends ZUtils implements InventoryManager {
     @Override
     public int getMaxPage(OfflinePlayer player) {
         return this.playerMaxPages.getOrDefault(player.getUniqueId(), 0);
+    }
+
+    @Override
+    public InventoryDefault getFakeInventory() {
+        InventoryDefault inventoryDefault = new InventoryDefault();
+        inventoryDefault.setPlugin(this.plugin);
+        return inventoryDefault;
     }
 }