+++ src/fr/maxlego08/menu/api/enums/ItemVerification.java
@@ -1,0 +1,8 @@
+package fr.maxlego08.menu.api.enums;
+
+public enum ItemVerification {
+
+    MODELID,
+    SIMILAR;
+
+}