+++ src/fr/maxlego08/menu/loader/MenuItemStackLoader.java
@@ -218,7 +218,9 @@ public class MenuItemStackLoader extends ZUtils implements Loader<MenuItemStack>
         if (item.getDisplayName() != null) configuration.set(path + "name", item.getDisplayName());
         if (!item.getLore().isEmpty()) configuration.set(path + "lore", item.getLore());
         if (item.isGlowing()) configuration.set(path + "glow", item.isGlowing());
-        if (item.getModelID() != null) configuration.set(path + "modelID", item.getModelID());
+        if (item.getModelID() != null && !item.getModelID().equalsIgnoreCase("0")) {
+            configuration.set(path + "modelID", item.getModelID());
+        }
         if (item.getData() > 0) configuration.set(path + "data", item.getData());
         if (item.getDurability() > 0) configuration.set(path + "durability", item.getDurability());
         if (item.getAmount() != null) {
@@ -290,7 +292,7 @@ public class MenuItemStackLoader extends ZUtils implements Loader<MenuItemStack>
         }
 
         if (item.getEnchantments() != null && !item.getEnchantments().isEmpty()) {
-            List<String> stringEnchantments = item.getEnchantments().entrySet().stream().map(e -> e.getKey().toString() + "," + e.getValue().toString()).collect(Collectors.toList());
+            List<String> stringEnchantments = item.getEnchantments().entrySet().stream().map(e -> e.getKey().getName() + "," + e.getValue().toString()).collect(Collectors.toList());
 
             configuration.set(path + "enchants", stringEnchantments);
         }